requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
dill==0.4.0
google-auth==2.42.1
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inquirerpy==0.3.4
multiprocess==0.70.14
orjson==3.11.0
pfzy==0.3.4
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
requests==2.32.5
robyn==0.72.2
rsa==4.9.1
rustimport==1.3.4
sniffio==1.3.1
tenacity==9.1.2
toml==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvloop==0.21.0
watchdog==4.0.1
wcwidth==0.2.14
websockets==15.0.1
